#63 Move files from utbot-gradle to utbot-framework module
