+++ utbot-framework/build.gradle
@@ -15,6 +15,7 @@ dependencies {
     fetchInstrumentationJar project(path: ':utbot-instrumentation', configuration:'instrumentationArchive')
 
     api project(':utbot-core')
+    api project(':utbot-summary')
     implementation 'junit:junit:4.13.1'
     api project(':utbot-framework-api')
