+++ utbot-gradle/src/main/kotlin/org/utbot/gradle/plugin/extension/SarifGradleExtensionProvider.kt
@@ -9,14 +9,7 @@ import org.utbot.framework.plugin.api.ClassId
 import org.utbot.framework.plugin.api.CodegenLanguage
 import org.utbot.framework.plugin.api.MockFramework
 import org.utbot.framework.plugin.api.MockStrategyApi
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.classesToMockAlwaysParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.codegenLanguageParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.forceStaticMockingParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.generationTimeoutParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.mockFrameworkParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.mockStrategyParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.staticsMockingParse
-import org.utbot.gradle.plugin.extension.SarifConfigurationParser.testFrameworkParse
+import org.utbot.framework.plugin.sarif.SarifExtensionProvider
 import java.io.File
 
 /**
