+++ utbot-maven/build.gradle
@@ -11,9 +11,7 @@ configurations {
 
 dependencies {
     // `compile` because `api` dependencies are not included in pom.xml by `install` task
-    compile project(':utbot-gradle')
     compile project(':utbot-framework')
-    compile project(':utbot-summary')
 
     implementation "org.apache.maven:maven-core:${maven_plugin_api_version}"
     implementation "org.apache.maven:maven-plugin-api:${maven_plugin_api_version}"
